feat(#172): the deck-fuel leg reads `expected`, not the pigeonhole floor
ADR-0076, four decisions. `_deck_basic_energy_fuel` answered "matching Basic
Energy a whole-deck search rider can PROVABLY still find" with the sound
pigeonhole floor, which collapses to ZERO on every realistic suite still
behind hidden prizes -- 3 unseen Water behind 5 prizes on 82756664 f97, a
deck 99.75% certain to hold Water. That zeroed min(recoverN, fuel, need) and
killed the accel dividend outright.

It now reads `CountTriple.expected` off the ONE `MySide.deck_energy_counts`
derivation. A ranked consumer asking HOW MANY takes the count leg; `p_any`
answers "is there at least one?" and weighting a full recoverN by it would
claim P(>=1) odds of finding ALL of them.

The `deck_known_counts` short-circuit is DELETED, not kept. It hid a real
divergence: the Pilot's hand-rolled prizes_hidden counts face-down entries,
and anchoring is our inference -- the engine never turns those cards face-up
-- so it read 5 forever while MySide.prizes_hidden drops to 0 on resolution.
The Count Triple collapses its own legs when anchored, so the regime branch
is subsumed rather than replaced.

Untyped riders (Turbo Flare, Whimsicott ex's Energy Gift -- "search your deck
for up to 3 Basic Energy cards") take the cross-type union, which is EXACT
rather than the second instrument ADR-0074 decision 6 forbids: every type's
leg divides the same (deck_count, prizes_hidden), so the sum over types IS
the aggregate.

Cost, stated rather than hidden:

- `_deck_basic_energy_fuel` / `_recover_units` / `_promote_accel_units` return
  float, and `PromoteBody.accel_units` becomes float. `my_yield` was coercing
  it with `int(...)`, which floored 2.5 to 2 and silently binned 37.5 points;
  two tests pin the fractional range and the sub-unit case.
- `_deck_basic_energy_fuel` drops its now-unused `board`/`obs` parameters and
  takes the StateModel instead.
- `test_turbo_flare_credit_is_need_and_fuel_gated` hand-injected
  `Board(deck_known_counts=...)` to reach the deleted branch. Rewritten to
  drive through the real `_board()`, preserving every behavioural claim it
  made (full credit with a benched Staryu, 0 firing blanks, 0 on a provably
  dry deck) and adding the #172 case.
- One stale docstring corrected: the refresh-supply estimate no longer
  contrasts itself against a pigeonhole floor no consumer takes.

Seams (all pre-existing, no new ones): test_state_model.py for the union
identity, test_promote_retreat_value.py for the fractional dividend,
test_promote_forward.py for the decision end-to-end plus the anchored
degeneracy guarantee.

Full suite green: 3928 passed, 1 skipped, 4 xfailed.

+++ src/common/promote_retreat_value.py
@@ -87,7 +87,9 @@ class PromoteBody:
     #: `_recover_units` count, need-gated. §3b: the dividend re-anchors on `ENERGY_RECOVER`, because
     #: retreating INTO Cinderace must credit what attacking WITH Cinderace credits (the shipped
     #: `_DIVIDEND = 5` was ~45x short of the 75-per-Energy the SAME rider earns on the attack).
-    accel_units: int = 0
+    #: FRACTIONAL: the deck-fuel leg is `CountTriple.expected` (ADR-0076), so this is an EXPECTED
+    #: count. `my_yield` must not coerce it — `int()` floored 2.5 to 2 and binned 37.5 points.
+    accel_units: float = 0.0
 
     # -- closure: the probabilistic middle (§5) -------------------------------------------------
     #: `max` over attacks of ``damage(a) x [readiness_p(a | enabler) - readiness_p(a)]`` — the
@@ -143,7 +145,7 @@ class PromoteBody:
         and its accel rider), which is the same axes discipline ADR-0069 §1 applies to the attach
         marginal — `max` WITHIN an axis, sum ACROSS."""
         attack = self.reach if self.wall_progress is None else float(self.wall_progress)
-        return max(0.0, attack) + ENERGY_RECOVER * max(0, int(self.accel_units))
+        return max(0.0, attack) + ENERGY_RECOVER * max(0.0, float(self.accel_units))
 
     def exposure(self) -> float:
         """The prizes promoting this body HANDS the opponent — continuous, per-body and
